+++ src/controllers/Input.php
@@ -1,11 +1,62 @@
 <?php
 
+/*
+The `Input::actionUpdateVisibleElements()` method is based on a large section of the
+`\craft\controllers\ElementsController::actionSaveDraft()` method, from Craft CMS 4.3.6.1,
+by Pixel & Tonic, Inc.
+https://github.com/craftcms/cms/blob/4.3.6.1/src/controllers/ElementsController.php#L1113
+Craft CMS is released under the terms of the Craft License, a copy of which is included below.
+https://github.com/craftcms/cms/blob/4.3.6.1/LICENSE.md
+
+Copyright © Pixel & Tonic
+
+Permission is hereby granted to any person obtaining a copy of this software
+(the “Software”) to use, copy, modify, merge, publish and/or distribute copies
+of the Software, and to permit persons to whom the Software is furnished to do
+so, subject to the following conditions:
+
+1. **Don’t plagiarize.** The above copyright notice and this license shall be
+   included in all copies or substantial portions of the Software.
+
+2. **Don’t use the same license on more than one project.** Each licensed copy
+   of the Software shall be actively installed in no more than one production
+   environment at a time.
+
+3. **Don’t mess with the licensing features.** Software features related to
+   licensing shall not be altered or circumvented in any way, including (but
+   not limited to) license validation, payment prompts, feature restrictions,
+   and update eligibility.
+
+4. **Pay up.** Payment shall be made immediately upon receipt of any notice,
+   prompt, reminder, or other message indicating that a payment is owed.
+
+5. **Follow the law.** All use of the Software shall not violate any applicable
+   law or regulation, nor infringe the rights of any other person or entity.
+
+Failure to comply with the foregoing conditions will automatically and
+immediately result in termination of the permission granted hereby. This
+license does not include any right to receive updates to the Software or
+technical support. Licensees bear all risk related to the quality and
+performance of the Software and any modifications made or obtained to it,
+including liability for actual and consequential harm, such as loss or
+corruption of data, and any necessary service, repair, or correction.
+
+THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES, OR OTHER
+LIABILITY, INCLUDING SPECIAL, INCIDENTAL AND CONSEQUENTIAL DAMAGES, WHETHER IN
+AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+
 namespace benf\neo\controllers;
 
 use benf\neo\elements\Block;
 use benf\neo\Plugin as Neo;
 use Craft;
 use craft\web\Controller;
+use craft\web\View;
 use yii\web\Response;
 
 /**
@@ -121,5 +172,112 @@ class Input extends Controller
         }
 
         return $return;
+    }
+
+    /**
+     * Gets data for which fields on a block should be visible.
+     *
+     * @return Response
+     * @since 3.7.0
+     */
+    public function actionUpdateVisibleElements(): Response
+    {
+        $this->requireAcceptsJson();
+        $this->requirePostRequest();
+
+        $request = Craft::$app->getRequest();
+        $elementsService = Craft::$app->getElements();
+        $fieldsService = Craft::$app->getFields();
+        $blocksData = $request->getRequiredBodyParam('blocks');
+        $fieldId = $request->getRequiredBodyParam('fieldId');
+        $ownerCanonicalId = $request->getRequiredBodyParam('ownerCanonicalId');
+        $ownerDraftId = $request->getRequiredBodyParam('ownerDraftId');
+        $siteId = $request->getRequiredBodyParam('siteId');
+        $sortOrder = $request->getRequiredBodyParam('sortOrder');
+        $field = $fieldsService->getFieldById($fieldId);
+        $canonicalOwner = $elementsService->getElementById($ownerCanonicalId, null, $siteId);
+        $blocks = $canonicalOwner::find()
+            ->provisionalDrafts()
+            ->draftId($ownerDraftId)
+            ->one()
+            ->getFieldValue($field->handle)
+            ->all();
+        $blockIds = array_map(fn($block) => $block->id, $blocks);
+        $returnedBlocksData = [];
+
+        foreach ($sortOrder as $i => $blockId) {
+            $blockData = $blocksData[$blockId];
+
+            if (is_int($blockId)) {
+                $block = Neo::$plugin->blocks->getBlockById((int)$blockId);
+            } else {
+                $block = Neo::$plugin->blocks->getBlockById((int)$blockIds[$i]);
+            }
+
+            if ($block === null) {
+                continue;
+            }
+
+            $view = Craft::$app->getView();
+            $namespace = "fields[{$field->handle}][blocks][{$block->id}]";
+            $fieldLayout = $block->getFieldLayout();
+            $form = $fieldLayout->createForm($block, false, [
+                'namespace' => $namespace,
+                'registerDeltas' => false,
+                'visibleElements' => $blockData['visibleLayoutElements'],
+            ]);
+            $missingElements = [];
+            foreach ($form->tabs as $tab) {
+                if (!$tab->getUid()) {
+                    continue;
+                }
+
+                $elementInfo = [];
+
+                foreach ($tab->elements as [$layoutElement, $isConditional, $elementHtml]) {
+                    /** @var FieldLayoutComponent $layoutElement */
+                    /** @var bool $isConditional */
+                    /** @var string|bool $elementHtml */
+                    if ($isConditional) {
+                        $elementInfo[] = [
+                            'uid' => $layoutElement->uid,
+                            'html' => $elementHtml,
+                        ];
+                    }
+                }
+
+                $missingElements[] = [
+                    'uid' => $tab->getUid(),
+                    'id' => $tab->getId(),
+                    'elements' => $elementInfo,
+                ];
+            }
+
+            $tabs = $fieldLayout->getTabs();
+            if (count($tabs) > 1) {
+                $selectedTab = isset($blockData['selectedTab']) && isset($tabs[$blockData['selectedTab']])
+                    ? $blockData['selectedTab']
+                    : null;
+                $tabHtml = $view->namespaceInputs(fn() => $view->renderTemplate('neo/_tabs.twig', [
+                    'tabs' => $tabs,
+                    'selectedTab' => $selectedTab,
+                    'block' => $block,
+                ], View::TEMPLATE_MODE_CP), $namespace);
+            } else {
+                $tabHtml = null;
+            }
+
+            $returnedBlocksData[$blockId] = [
+                'tabs' => $tabHtml,
+                'missingElements' => $missingElements,
+                'initialDeltaValues' => $view->getInitialDeltaValues(),
+                'headHtml' => $view->getHeadHtml(),
+                'bodyHtml' => $view->getBodyHtml(),
+            ];
+        }
+
+        return $this->asSuccess(data: [
+            'blocks' => $returnedBlocksData,
+        ]);
     }
 }